Revert "Update CalendarUI"
